run percy on all pull-requests
this changes the action to no longer run when a 'percy' label is added to a pull-request and instead to run on all pull-requests.

the reason we had the label trigger the action was because we had a very small allowance of percy runs available and we wanted to only run percy when we thought we needed to.

We now have a much higher allowance of percy runs, enough to be able to run the action on all pull-requests. This will remove the need for someone to remember to add a percy label to a pull-request which changes the ui.